chore(api): remove constants file
