+++ src/utils/dataSetReport/isCustomFormType.js
@@ -1,0 +1,1 @@
+export const isCustomFormType = formType => formType === 'CUSTOM'
